feat(auth): add verifyToken utility
